+++ app/Http/Controllers/ItemController.php
@@ -3,7 +3,7 @@
 namespace App\Http\Controllers;
 
 use App\Http\Parameters\Get;
-use App\Http\Parameters\Route\Validate;
+use App\Validators\Request\Route;
 use App\Models\Category;
 use App\Models\Item;
 use App\Models\SubCategory;
@@ -41,7 +41,7 @@ class ItemController extends Controller
      */
     public function index(Request $request, string $resource_type_id, string $resource_id): JsonResponse
     {
-        Validate::resourceRoute($resource_type_id, $resource_id);
+        Route::resourceRoute($resource_type_id, $resource_id);
 
         $this->collection_parameters = Get::parameters([
             'include-categories',
@@ -109,7 +109,7 @@ class ItemController extends Controller
         string $item_id
     ): JsonResponse
     {
-        Validate::itemRoute($resource_type_id, $resource_id, $item_id);
+        Route::itemRoute($resource_type_id, $resource_id, $item_id);
 
         $item = (new Item())->single($resource_type_id, $resource_id, $item_id);
 
@@ -137,7 +137,7 @@ class ItemController extends Controller
      */
     public function optionsIndex(Request $request, string $resource_type_id, string $resource_id): JsonResponse
     {
-        Validate::resourceRoute($resource_type_id, $resource_id);
+        Route::resourceRoute($resource_type_id, $resource_id);
 
         $this->collection_parameters = Get::parameters(['year', 'month', 'category', 'subcategory']);
 
@@ -178,7 +178,7 @@ class ItemController extends Controller
         string $item_id
     ): JsonResponse
     {
-        Validate::itemRoute($resource_type_id, $resource_id, $item_id);
+        Route::itemRoute($resource_type_id, $resource_id, $item_id);
 
         $item = (new Item())->single($resource_type_id, $resource_id, $item_id);
 
@@ -217,7 +217,7 @@ class ItemController extends Controller
      */
     public function create(Request $request, string $resource_type_id, string $resource_id): JsonResponse
     {
-        Validate::resourceRoute($resource_type_id, $resource_id);
+        Route::resourceRoute($resource_type_id, $resource_id);
 
         $validator = (new ItemValidator)->create($request);
 
@@ -267,7 +267,7 @@ class ItemController extends Controller
         string $item_id
     ): JsonResponse
     {
-        Validate::itemRoute($resource_type_id, $resource_id, $item_id);
+        Route::itemRoute($resource_type_id, $resource_id, $item_id);
 
         if ($this->isThereAnythingToPatchInRequest() === false) {
             UtilityResponse::nothingToPatch();
@@ -330,7 +330,7 @@ class ItemController extends Controller
         string $item_id
     ): JsonResponse
     {
-        Validate::resourceRoute($resource_type_id, $resource_id);
+        Route::resourceRoute($resource_type_id, $resource_id);
 
         $item = (new Item())->single($resource_type_id, $resource_id, $item_id);
